// apps/chatbots/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
// ...
  public: {
    Tables: {
// ...
      chatbots: {
        Row: {
          account_id: string;
          created_at: string;
          description: string | null;
          id: string;
          name: string;
          settings: Json;
          site_name: string;
          url: string;
        };
        Insert: {
          account_id: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          name: string;
          settings?: Json;
          site_name: string;
          url: string;
        };
        Update: {
          account_id?: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          name?: string;
          settings?: Json;
          site_name?: string;
          url?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'chatbots_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'accounts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'chatbots_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_account_workspace';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'chatbots_account_id_fkey';
            columns: ['account_id'];
            isOneToOne: false;
            referencedRelation: 'user_accounts';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      documents: {
        Row: {
          chatbot_id: string;
          content: string;
          created_at: string;
          hash: string;
          id: string;
          title: string;
        };
        Insert: {
          chatbot_id: string;
          content: string;
          created_at?: string;
          hash: string;
          id?: string;
          title: string;
        };
        Update: {
          chatbot_id?: string;
          content?: string;
          created_at?: string;
          hash?: string;
          id?: string;
          title?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'documents_chatbot_id_fkey';
            columns: ['chatbot_id'];
            isOneToOne: false;
            referencedRelation: 'chatbots';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
// ...
};